+++ apps/client/src/app/modules/gearsets/gearsets.module.ts
@@ -4,18 +4,19 @@ import { StoreModule } from '@ngrx/store';
 import { EffectsModule } from '@ngrx/effects';
 import * as fromGearsets from './+state/gearsets.reducer';
 import { GearsetsEffects } from './+state/gearsets.effects';
-import {
-  NzAlertModule,
-  NzButtonModule,
-  NzDividerModule,
-  NzFormModule,
-  NzIconModule,
-  NzInputModule,
-  NzMessageModule, NzModalModule,
-  NzPopconfirmModule,
-  NzSelectModule, NzTagModule, NzTimelineModule,
-  NzToolTipModule
-} from 'ng-zorro-antd';
+import { NzAlertModule } from 'ng-zorro-antd/alert';
+import { NzButtonModule } from 'ng-zorro-antd/button';
+import { NzDividerModule } from 'ng-zorro-antd/divider';
+import { NzFormModule } from 'ng-zorro-antd/form';
+import { NzIconModule } from 'ng-zorro-antd/icon';
+import { NzInputModule } from 'ng-zorro-antd/input';
+import { NzMessageModule } from 'ng-zorro-antd/message';
+import { NzModalModule } from 'ng-zorro-antd/modal';
+import { NzPopconfirmModule } from 'ng-zorro-antd/popconfirm';
+import { NzSelectModule } from 'ng-zorro-antd/select';
+import { NzTagModule } from 'ng-zorro-antd/tag';
+import { NzTimelineModule } from 'ng-zorro-antd/timeline';
+import { NzToolTipModule } from 'ng-zorro-antd/tooltip';
 import { TranslateModule } from '@ngx-translate/core';
 import { GearsetCreationPopupComponent } from './gearset-creation-popup/gearset-creation-popup.component';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';