+++ apps/client/src/app/modules/number-question-popup/number-question-popup/number-question-popup.component.ts
@@ -1,5 +1,5 @@
 import { Component, Input } from '@angular/core';
-import { NzModalRef } from 'ng-zorro-antd';
+import { NzModalRef } from 'ng-zorro-antd/modal';
 
 @Component({
   selector: 'app-name-question-popup',