+++ apps/client/src/app/modules/user-picker/user-picker/user-picker.component.ts
@@ -2,7 +2,7 @@ import { Component } from '@angular/core';
 import { combineLatest, Observable, of } from 'rxjs';
 import { FormControl, Validators } from '@angular/forms';
 import { CharacterSearchResult, CharacterSearchResultRow, XivapiService } from '@xivapi/angular-client';
-import { NzModalRef } from 'ng-zorro-antd';
+import { NzModalRef } from 'ng-zorro-antd/modal';
 import { debounceTime, map, shareReplay, switchMap, tap } from 'rxjs/operators';
 import { UserService } from '../../../core/database/user.service';
 import { AuthFacade } from '../../../+state/auth.facade';