+++ apps/client/src/environments/game-env.ts
@@ -1,6 +1,6 @@
 export const gameEnv = {
   maxLevel: 80,
-  globalGameVersion: 5.35,
+  globalGameVersion: 5.351,
   koreanGameVersion: 5.21,
   chineseGameVersion: 5.25
 };